CR 3550391: Revert "Debian: Make assembly code PIC-compatible"
This reverts commit 832bfdd7a52530895a726b48956db2326cf0fcbc.

The Debian change was intended to fix build breakage when building
shared libraries without -Bsymbolic (and when using assembly packs).
However, it seems that upstream commit
ddc20d4da9d770afeace2b4111d5d39d00e6c6b4 has provided a work-around
for the build breakage for 5 years, and Debian did not update to remove
their old patch.

In any case, Akamai is building with -Bsymbolic via the defaults in
Makefile.shared, so we would not be affected by the issue regardless.

To be squashed into commit 832bfdd7a52530895a726b48956db2326cf0fcbc
(so that both are removed) on rebase.