CR 3605931: set OPENSSL-TARGET for non-MASTER builds
As part of our openssl "componentization", we set $P/OPENSSL-TARGET
to build_crypto so that only libcrypto is compiled, and libssl
does not even exist unless pulled in by openssl/enable/tlsv1 (or similar).
That change was lost in the addition of shared library support in
7.102.x, resulting in a functioning libssl being present when it
was not expected to be (i.e., for static builds that did not
pull in openssl/enable/tlsv1).

Set $P/OPENSSL-TARGET to build_crypto by default as appropriate,
to restore the expected behavior.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal