CR 3598201: give akamake OS X builds a configure target
For the shared-library capable versions, we split up the
configure step into two rules, one for static and one for shared,
and pick between them depending on the desired openssl linkage.
We took that opportunity to switch from calling config to calling
Configure in the akamake rules, since we were explicitly passing
a configure target on linux systems, and that was appended to the
one that config picked, which caused Configure to get confused.
However, that left osx builds from akamake with no configure targets,
though we did have a separate debug case.

Add another else if to the long chain in akamake-env.mak to let
non-debug os x builds from akamake succeed.