CR 3548491: SSLv2 crashes
The code that changes the default set of ciphers was attemping to
modify the list of ssl2_ciphers indirectly by going though ALL ciphers.

The ssl2_cipher list was declared const, so it was in read-only memory
causing a crash when attempting to write to it. Modify the table to be
non-const.

Original Author: tshort@akamai.com
Code Reviewer: pbohman@akamai.com
Akamai Status: internal