CR 3611591: Use _WIN32 over WIN32 for preprocessor
Though upstream OpenSSL does use both the with- and without-underscore
versions of this preprocessor symbol in conditionals, the intent seems
to be that the WIN32 symbol is a direct byproduct of being a
windows-variant configuration and should be used for feature
en/disablement on windows systems.  Use of the _WIN32 symbol is
more widespread, being used to implement platform portability of
more generic code.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: to be submitted (seed.c is the only part that applies)