CR 3611591: Let testharness.c build on windows
Much of this can be had by just including e_os.h.
The internal-only header e_os.h provides some OS portability
constructs such as defining stderr to a useful value on windows.
The functionality of that header includes namespace and other pollution
not appropriate for a library to leak into its consumers, but is useful
for internal code that we have added.

We also need to not include Unix-only headers unistd.h and pthread.h,
which are apparently not needed to build here; crib the conditional from
mem_sec.c to only include them on not-windows.

Finally, windows does not provide a usleep() function, but since we
are only sleeping in the milliseconds regime, the Sleep() function
will work fine.

This should help the Akamai fork of OpenSSL 1.1.0 to build on windows
systems (even via normal Configure/nmake not using akamake).

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal