CR 3561461: Move CLIENTHELLO_MSG up in the header
We'll be adding a field of this type to struct ssl_st in a subsequent
commit, and need the type definition to be in scope already.
Also move up the RAW_EXTENSION definition that it depends on.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: merged in master