CR 3644751: remove too-strict symbol version check
Upstream has a check that looks at the length of the version string
in an attempt to check that a symbol is not added at an earlier letter
version than already exists, which is inconsistent with their release
strategy.

However, we are forced to have the appearance of doing so (adding
a symbol at an earlier letter version than already exists), since we
have to add symbols at artificially large numbers in order to avoid
conflicting with upstream's numbers.  That is, our symbols are all at
the end, and not interleaved in the actual order that they are added.

So, it is normal for the transition from upstream symbols to akamai
symbols to go from (e.g.) 1_1_0d to 1_1_0, which this length check
detects.  Remove the check that does not apply to our fork.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal

(cherry picked from commit 909d0ef1eb2963c151aba02de498a1278967e217)