CR 3650493: add SSL_akamai_remove_session()
Sometimes decides that a session being resumed forward to origin is bad
and needs to be regenerated.  It used to do so by calling
SSL_set_session(s, NULL) and clearing s->hit (a little removed
from each other in the codepath), but cannot access s->hit with
the opaque data structures in openssl 7.110.x.

Provide a helper function to perform the consolidated operation
and remove the need for the write accessor.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: to be published