CR 3640401: Copy openssl shared libs to common/lib
Copy the openssl shared libraries to common/lib
This really adds the .so.102.0.0 files.
Do not copy the libopenssl.so library, it's just a copy of
libssl.so, and no one is supposed to be linking against that
anymore; if something were to try to link against libopenssl.so,
they'd get a reference to libssl.so.