CR 3655303: define SSL_AKAMAKI_TLSEXT_ERR_IGNORE
Provide a value that an ALPN callback can return so as to obtain
the same library behavior that would occur if no callback was
registered.  Due to upstream commit 062178678f5374b09f00d70796f6e692e8775aca,
the presence of such a callback is taken as a proxy to indicate that
the application has configured ALPN protocols, and as such, that any
non-success return from the callback is either a hard failure or indicates
no overlap between the client and server protocol list, which is
a SHALL-level requirement to fail, in RFC 7301.

But, having a callback set need not indicate that the application is
actually intending to configure ALPN for this connection, so, allow
a way to opt out of this behavior, while preserving the upstream
behavior for SSL_TLSEXT_ERR_NOACK.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: ask upstream