CR 3665672: blank lines in PEM body are errors
Bringing in the new tests revealed an unexpected behavior in the
post-refactor behavior of the PEM read pipeline, namely that we
accept blank lines in the base64 body of encrypted PEM files.
This was an error in the old implementation, and is not really
a good idea to allow anyway, so bring back the old behavior.

Despite looking a bit odd, PEM_R_BAD_END_LINE seems to be the reason
code that the old code used in this case, since all short lines were
treated as breaking out of the loop, and the next step was a check
for the "-----END ...-----" line, which fails.

To be squashed on rebase.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: to-be published

(cherry picked from commit 91b21f9ee93963328e7acf85553a7f9d8c15dd25)