CR 3694412: preserve ClientHello extension order
We need to sort the received extensions in order to efficiently detect
duplicates, which are forbidden by the spec.  But we also want to
process the extensions in the order supplied by the client, so as to
be able to use that information for fingerprinting clients.
So, do dupliate detection on a copy instead of in-place.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: proprietary (not relevant on master)