CR 3714182: rename PEM flag to match upstream
PEM_FLAG_WEAK_EOL --> PEM_FLAG_EAY_COMPATIBLE is slightly more clear
about why it is doing what it is doing.

Upstream also has a PEM_FLAG_ONLY_B64, but we do not have any consumers
of it, so there does not seem to be a need to pull it in early.