CR 3726352: disable afalgeng and capieng for openssl-sgx
The AFALG engine is only built if configure detects that the kernel
available is new enough to support the required calls.  As of alsi9,
it is, which causes trouble for the limited environment of openssl-sgx.
Disable it unconditionally instead of relying on the configure check
to get the desired behavior.

Also disable capieng while here; though it's for Windows APIs, a
similar principle holds.