CR 3742982, 3803901: Akamai option to disable EMS per-connection
CR 3742982: Akamai option to disable EMS per-connection

Add SSL_AKAMAI_OPT_NO_EMS, that can be set on either an SSL or an
SSL_CTX.  When processing a ClientHello, if this flag is set, do
not indicate that the EMS TLS extension was received in either the
ssl3 object or the SSL_SESSION.  Retain most of the sanity checks
between the previous and current session during session resumption,
but weaken the check when the current SSL object is configured to
not use EMS.

Interestingly, there seem to be two places in the code that parse
the ClientHello and set the TLS1_FLAGS_RECEIVED_EXTMS flag.  We
modify both locations, but there is probably some general cleanup
waiting to happen.

Test case for normal no-EMS operation

That is, just make sure that a client with EMS can talk to a server
with it disabled, and both report that it's not available.

Does not test resumption behavior at all.

CR 3803901: SSL_AKAMAI_OPT_NO_EXTMS for client SSLs

The SSL_AKAMAI_OPT_NO_EXTMS option was added for use on server SSL/SSL_CTXes,
needed when cryptoserver is in use, as the ghost/cryptoserver protocol does
not convey the needed information for the full handshake transcript hash
that is used as input to the extended master secret calculation.

We also need the ability to prevent the extended master secret extension
from being sent in outgoing ClientHellos, to allow connecting to remote
servers that misbehave when presented with such extensions (as generated
by OpenSSL), which appear as empty extensions at the end of the ClientHello.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: https://github.com/openssl/openssl/pull/3910