CR 3742982: Akamai option to disable EMS per-connection
Add SSL_AKAMAI_OPT_NO_EMS, that can be set on either an SSL or an
SSL_CTX.  When processing a ClientHello, if this flag is set, do
not indicate that the EMS TLS extension was received in either the
ssl3 object or the SSL_SESSION.  Retain most of the sanity checks
between the previous and current session during session resumption,
but weaken the check when the current SSL object is configured to
not use EMS.

Interestingly, there seem to be two places in the code that parse
the ClientHello and set the TLS1_FLAGS_RECEIVED_EXTMS flag.  We
modify both locations, but there is probably some general cleanup
waiting to happen.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: proprietary