CR 3745822: only set --strict-warnings on linux
Move the default setting out of Configure and into akamake-env.mak,
applying only to the akamai-linux targets (not akamai-darwin or windows),
as those are the ones that get built nightly in BaRT and the BaRT
builds are the main ones used.  Windows does not build in BaRT, and darwin
is used on developer workstations, where the environment differs greatly
from the BaRT-provided environment.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: proprietary