CR 3736922: consolidate shared session cache and unify locking
Previously, we stored a session list structure in SSL_CTX ex_data
that held head and tail pointers for the list linkage of SSL_SESSION
objects; these lists could be shared across multiple SSL_CTX objects
and its lock/reference count used to protect the lhash of sessions
that remained in each SSL_CTX.  When assigning SSL_CTXes to share a
session cache, plain pointer aliasing was used for the lhash structure,
which led to inconsistent application of the requirement to hold the
lock from the shared session_list object across accesses to it.

By moving the lhash structure to the shared session_list object,
the ownership and locking model are made more clear.

The locking is updated to hold the session_list lock across all accesses
to the list head/tail pointers, the session lhash, and the list
structure; this includes several fixes for previously unprotected
accesses.  Unfortunately, we cannot properly lock an SSL_CTX_ctrl()
function used to get the number of sessions in the cache (via the
lhash object), but this is deemed to be acceptabls since such information
is currently only used to determine when to evict stale entries and
maintain a cap on the size of the session cache; being somewhat flexible
in the application of that cap should not have far-reaching consequences.

Provide a convenience API to lock the session list, search the lhash for
a matching session, take a reference on that session, and return it,
so that intrusive locking changes are not needed at most call sites.

We furthermore make the SSL_CTX_sessions() API useless, as the lock
for the shared session cache is not exposed to library consumers;
reflect this reality by always returning failure from that function.
Provide an alternative for ghost's use that combines two operations
used in printing statistics.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: to-be-published (related to github issue #2397)