CR 3827801: extensively limit akamake via ATTR_ONLY
Akamake does not commit to making BUILDENV determinations for ATTR_ONLY
builds, so we are essentially asking for undefined behavior when we
unconditionally attempt to inspect the BUILDENV and enforce sanity/consistency
checks on the shared/static linkage of the library.  Additionally, when
users manually specify (e.g.) BUILDENV.BASEOS and ATTR_ONLY on the command
line so as to be able to build the dep_comp target outside of a VBE,
the code prior to this commit would trigger the big scary error message
about needing to choose a linkage.

There's no need to enforce the linkage decision for ATTR_ONLY builds,
so apply the appropriate conditional and limit it to "real" builds.

Going further, when ATTR_ONLY is in effect, the only interface we
really need to provide is dep_comp-related, so in that case, limit ourselves
to akamake boilerplate, clean targets, and setting our DEP-COMPS.

To be squashed into other Akamake changes on rebase.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal