CR 3846451: make versioned .so's available during the build
In CR 3821801 we reworked how the library export is done and removed
a pair of dependencies that seemed unneeded, from the perspective
of having the versioned .so's appear in common/lib by the end of the
build.  However, the only means within Akamake that forces EXP-LIBS
to appear in common/lib are via an explicit request for a DEP-LIBS
(that produces a lib.a.dep file and the corresponding lib.a), and
via a dependency from build-galaxy on $(call libs-available, .).
The DEP-LIBS path suffices for link-time, which uses the lib.a path,
but is insufficient for binaries that are being run at build time,
which requires the versioned .so path (as well as our LD_LIBRARY_PATH
shenanigans).

Because we do not have a way to know when the shared library will
be used for execution during a build, and to provide parity with
the behavior of staticly linked builds, whenever a specific library
is pulled in for linking, ensure that the versioned .so is also
available for use at that time.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal