CR 3853701: use correct version in versioned .so paths
The previous commit was cherry-picked from the 7.110.6-dev branch,
which made it easy to not notice that the .110.0.0 SOVERSION was still
in the dependency line.  This branch uses a SOVERSION with .111.0.0,
so we need to update the dependency line to match the file that is
actually going to be produced.

To be squashed into commit ad613ce28835c070cf70933d749ae5f124def657 on
rebase.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: internal