CR 3864571: Fix SSL_writev()
Improve performance of SSL_writev() by removing the memory allocation
and copy. As with SSL_readv(), reference the buckets in ex_data and
use them deep inside the write function where the data is copied from
the input buffer; this does not introduce an additional copy, the copy
is there to handle compression. If there were no compression support,
then the write code could be updated a la the "multiblock" code, and
encrypt directly from the input buffer. That would make SSL_writev()
much harder to implement.

Thus, SSL_writev() does not work well with compression (which can lead
to the CRIME attack, so probably a good thing); a slow path (alloc/copy)
is followed.

SSL_writev() also does not work with the "multiblock" functionality
due to reading directly from the input buffer. This is not as bad as
it sounds, since the "multiblock" functionality is limited to TLSv1.1
and TLSv1.2 with AES-CBC-SHA1/256 ciphers and no compression.

Also, no DTLS support.