CR 3868691: accept supported_groups in ServerHello
Mark the "supported_groups" (formerly "elliptic_curves") extension
as permitted in TLS 1.2 (and below) ServerHello messages.  Some
servers erroneously generate this extension in this message,
which was accepted by OpenSSL 1.1.0 and previous.  To avoid
a regression of functionality, we need to relax our protocol
compliance checks for this specific case.

Since we continue to not specify a tls_parse_stoc_supported_groups
function, there is no logic that needs updating; the extensions
framework will cope with this scenario as-is.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: to-be-submitted