CR 3890071: CID 39710 deref before null check
We ensure at SSL_CTX creation time that the session_list in ex_data
is non-NULL, so that can be treated as an invariant for code
not involved in SSL_CTX initialization.

To be squashed on rebase.

Original Author: bkaduk@akamai.com
Code Reviewer: tshort@akamai.com
Akamai Status: proprietary