CR 3919591: avoid hypothetical double-hash in RSALG+TLS1.3
RSALG works by using hash output as the server random and
the unhashed input as an authenticator to the cryptoserver
to authenticate the signing operation; the unhashed input can
then be discarded to avoid later replays.  TLS 1.3 draft-22
control flow has us send two ServerHello messages (one of
which acts as a HelloRetryRequest, with a magic value in
the 'random' field), so we "need to" be careful to not
hash the random value twice, once for each ServerHello message.

This is a non-issue in practice, since the two-ServerHello
flow only happens for TLS 1.3, and cryptoserver is
incompatible with TLS 1.3 (and extended master secret, for
similar reasons), but we can still make the code more
clear about what is happening.

Original Author: bkaduk@akamai.com
Code reviewer: tshort@akamai.com
Akamai Status: internal