CR 3952951: ABI compatibility 7.110 to 7.111
* Map SSL_AKAMAI_OPT_DISALLOW_RENEGOTIATION to OpenSSL SSL_OP_NO_RENEGOTIATION option.
* Add 7.110-named wrapper functions for functions added upstream.
* Add previous akamai name for ASN1_TIME_cmp_time_t()
* Add SSL_akamai_get_tmp_cipher() support
* Add some SSL_INTERNAL functions so that we can pass 1.1.0 unit tests
The functions are no-ops, and are deprecated in 1.2

Fix versions of functions in .num

We added PEM_read_bio_ex() to 1.1.0e, so we need to indicate that in the
libcrypto.num linkage file; otherwise, it won't be found by 1.1.0
compiled programs.

We added SSL_CTX_set1_cert_storex(), SSL_CTX_user_cert_and_key(), and
SSL_use_cert_and_key() to 1.1.0, so we need to indicate that in the
libssl.num linkage file; otherwise, it won't be found by 1.1.0 compiled
programs.

The ordinals don't matter, but the version does, since it's translated to:

    <function@>OPENSSL_<version>

in the shared object.

Original Author: tshort@akamai.com
Code Reviewer: bkaduk@akamai.com
Akamai Status: internal only