CR 3970121: compatibility with out-of-tree builds
The initial import of openssl-sgx changed the Configure script
to add a forced compiler include of
crypto/include/internal/bypass_to_sgxssl.h, as part of forcing SGX use.

This is not compatible with the nascent out-of-tree builds feature in
Akamake 5, that will let third-party software use a separate object
directory from source directory, to keep a clean separation and
facilitate cleanup steps.  This is because the referenced path
does not exist in the separate object tree, and we instead need to
explicitly refer to the source tree.  For most things, make's vpath
support will silently splice in the source tree for us, but not
in this case.