CR 4079861: Fix OPENSSL-TARGET for parallel builds
OPENSSL-TARGET was potentially getting long, and causing issues with
parallel builds (specifically two targets that build the same thing).
So, assign to just "all" if thats the case, or if setting to
"libcrypto.a", check to see if "all" is already in the list.

Original Author: tshort@akamai.com
Code Review: bkaduk@akamai.com
Akamai Status: internal only