CR 4292165: DSO is no longer optional
As of upstream commit 5fba3afad01707f4a8856a35500de007a8a256ec, DSO
support is enabled by default and no-dso is supposed to be ignored
(it gets an entry in %deprecated_disablables and is silently accepted).
However, our Akamai patches to the default configuration still were
setting "dso" as disabled, which means that we do not descend into
crypto/dso, and thus do not build the implementations of the DSO
functions that everything else needs to link.

Drop "dso" from %disabled, make enable/dso/akamake a silent no-op, and
do not try to pass enable-dso for shared builds.