CR 4287005: Fix new cache_id_changes
When using the legacy API (IP addrs), set the size based on IP address type.
The sockaddr_in(6) structures are padded out to sockaddr_storage and
sockadd_in6 is non-contiguous. Copying the structures included potentially
random data, which would prevent the matching. So, zero out the non-used
fields so they can be matched.